pref: PageResult 的 current、size 调整为 long 类型

+++ helio-core/src/main/java/cc/uncarbon/framework/core/page/PageResult.java
@@ -2,14 +2,15 @@ package cc.uncarbon.framework.core.page;
 
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
-import java.io.Serializable;
-import java.util.Collections;
-import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.Accessors;
 import lombok.experimental.SuperBuilder;
+
+import java.io.Serializable;
+import java.util.Collections;
+import java.util.List;
 
 /**
  * 分页查询结果
@@ -24,10 +25,10 @@ import lombok.experimental.SuperBuilder;
 public class PageResult<T extends Serializable> implements Serializable {
 
     @ApiModelProperty(value = "当前页")
-    private int current;
+    private long current;
 
     @ApiModelProperty(value = "当前页数量")
-    private int size;
+    private long size;
 
     @ApiModelProperty(value = "总量")
     private long total;
